Moved TestUtils and EnqueueMockResponse from src/test/java to the src/main/java to make it easier to share them with other modules.
