+++ sdk/core/azure-core/build.gradle
@@ -19,7 +19,7 @@ dependencies {
     implementation "com.squareup.okhttp3:okhttp:$okHttpVersion"
     implementation "com.squareup.retrofit2:retrofit:$retrofitVersion"
     implementation "javax.xml.stream:stax-api:$staxApiVersion" // https://stackoverflow.com/a/47371517/1473510
-    testImplementation "com.squareup.okhttp3:mockwebserver:$okHttpVersion"
+    implementation "com.squareup.okhttp3:mockwebserver:$okHttpVersion"
     testImplementation "junit:junit:$jUnitVersion"
     testImplementation "org.threeten:threetenbp:$threeTenBpVersion"
 }
