+++ sdk/core/azure-core/src/main/java/com/azure/android/core/internal/util/test/TestUtils.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.core.http.interceptor;
+package com.azure.android.core.internal.util.test;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
