+++ sdk/core/azure-core/src/test/java/com/azure/android/core/http/interceptor/CurlLoggingInterceptorTest.java
@@ -1,5 +1,7 @@
 package com.azure.android.core.http.interceptor;
 
+import com.azure.android.core.internal.util.test.EnqueueMockResponse;
+
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -14,9 +16,9 @@ import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.mockwebserver.MockWebServer;
 
-import static com.azure.android.core.http.interceptor.TestUtils.buildOkHttpClientWithInterceptor;
-import static com.azure.android.core.http.interceptor.TestUtils.getSimpleRequest;
-import static com.azure.android.core.http.interceptor.TestUtils.getSimpleRequestWithHeader;
+import static com.azure.android.core.internal.util.test.TestUtils.buildOkHttpClientWithInterceptor;
+import static com.azure.android.core.internal.util.test.TestUtils.getSimpleRequest;
+import static com.azure.android.core.internal.util.test.TestUtils.getSimpleRequestWithHeader;
 import static com.azure.android.core.util.logging.ClientLogger.LOG_LEVEL_DEBUG;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
