+++ sdk/storage/azure-storage-blob/build.gradle
@@ -17,7 +17,6 @@ android {
 
     sourceSets {
         androidTest.assets.srcDirs += files("$projectDir/schema".toString())
-        test.java.srcDirs += ["${project(':sdk:core:azure-core').projectDir}/src/test/java"]
     }
 }
 
@@ -36,6 +35,5 @@ dependencies {
     implementation "javax.xml.stream:stax-api:$staxApiVersion" // https://stackoverflow.com/a/47371517/1473510
     testImplementation "com.squareup.okhttp3:mockwebserver:$okHttpVersion"
     testImplementation "junit:junit:$jUnitVersion"
-    testImplementation "org.threeten:threetenbp:$threeTenBpVersion"
     annotationProcessor "androidx.room:room-compiler:$roomVersion"
 }
