+++ sdk/storage/azure-storage-blob/src/test/java/com/azure/android/storage/blob/interceptor/ResponseHeadersValidationInterceptorTest.java
@@ -13,7 +13,7 @@ import okhttp3.Request;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 
-import static com.azure.android.core.http.interceptor.TestUtils.buildOkHttpClientWithInterceptor;
+import static com.azure.android.core.internal.util.test.TestUtils.buildOkHttpClientWithInterceptor;
 
 public class ResponseHeadersValidationInterceptorTest {
     private static final String CLIENT_ID_HEADER = "x-ms-client-id";
